Refactor server routes for modular file-based routing Refactor file structure to hide api keys and ignore api key files on git
