+++ .gitignore
@@ -2,6 +2,9 @@ node_modules/
 bower_components/
 *.log
 
+## api keys ##
+server/config/
+
 build/
 dist/
 public/compiled/
