+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "compile": "babel . --out-dir public/compiled --presets=es2015,react --ignore=node_modules,compiled,server,db --source-maps inline --watch"
+    "compile": "babel . --out-dir public/compiled --presets=es2015,react --ignore=node_modules,compiled,server,db --source-maps inline --watch",
+    "start": "nodemon ./server/server.js"
   },
   "repository": {
     "type": "git",
