+++ server/cron.js
@@ -8,14 +8,16 @@
 var CronJob = require('cron').CronJob;
 const db = require('../db');
 const dbModels = require('../db/index.js');
+const twilioConfig = require('./config/twilio.js')
 
 //I will delete this account information after completion of our Greenfield project, 
 //but you can sign up for a Twilio account here with some amount of free credit: https://www.twilio.com/
-var accountSid = 'AC4d6654e30b96d8216b289b5cdcc918e9'; 
-var authToken = '6f2e5c9ffa8f5d3ba10438c4d0b4d9bf';
+// var accountSid = 'AC4d6654e30b96d8216b289b5cdcc918e9'; 
+// var authToken = '6f2e5c9ffa8f5d3ba10438c4d0b4d9bf';
 
 //Twilio REST API: http://twilio.github.io/twilio-node/
-var client = require('twilio')(accountSid, authToken); 
+// var client = require('twilio')(accountSid, authToken); 
+var client = require('twilio')(twilioConfig.accountId, twilioConfig.token);
 
 var job = new CronJob('10 * * * * *', 
   function() {
